BTA-1728 validate webhook request headers for php (#104)
* BTA-1728 Add request validation to the PHP SDK using the Authorization Signature in the request headers.

The primary use case for this is to validate webhooks

The

* BTA-1728 Move the validate request method outside the loop

The method to validate the request signature was being generated for every API method.
This commit moves the method outside the api operations loop so it's done just once per api class.

* BTA-1728 Update webhook header validation method with API key check

This change ensures that the API key in the webhook request headers matches the configured API key while validating the signature.
It also adds a guard clause to the README to only parse the webhooks when validation passes.

* BTA-1728 Modify test harness to handle API keys with spaces and special characters