BTA-1674-Add-Ability-In-Our-SDK-To-Parse-And-Validate-Webhook-Headers-Java (#109)
Added method that verifies the webhooks headers signature before they are parsed.
Extracted code from the `buildRequest` method that creates the `Authorization-Signature`
when the request is built and put it into a new method called`signRequest`, now we are
able to re-use said code when verifying webhooks. Made minor text correction to the ruby
template.